// src/clarity-angular/wigfrid/components/FlexGrid/Grid/enum/CellType.d.ts
/**
 * Identifies the type of cell in a @see:GridPanel.
 */
export declare enum CellType {
    /** Unknown or invalid cell type. */
    None = 0,
    /** Regular data cell. */
    Cell = 1,
    /** Column header cell. */
    ColumnHeader = 2,
    /** Row header cell. */
    RowHeader = 3,
    /** Top-left cell. */
    TopLeft = 4,
}