// src/clarity-angular/wigfrid/components/FlexGrid/Grid/enum/AllowDragging.d.ts
/**
 * Specifies constants that define the row/column dragging behavior.
 */
export declare enum AllowDragging {
    /** The user may not drag rows or columns. */
    None = 0,
    /** The user may drag columns. */
    Columns = 1,
    /** The user may drag rows. */
    Rows = 2,
    /** The user may drag rows and columns. */
    Both = 3,
}