// src/clarity-angular/wigfrid/components/FlexGrid/Grid/enum/AllowResizing.d.ts
/**
 * Specifies constants that define the row/column sizing behavior.
 */
export declare enum AllowResizing {
    /** The user may not resize rows or columns. */
    None = 0,
    /** The user may resize columns. */
    Columns = 1,
    /** The user may resize rows. */
    Rows = 2,
    /** The user may resize rows and columns. */
    Both = 3,
}